Much Cleaning
Replacement policies are now created through a factory (so files don't
need to be required all over the place and so that the replacement
policies don't need special case logic for direct mapped caches).
Moved the profiler script out of lib - it's the next thing on the list
of stuff to clean up.